requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
alembic==1.13.1
bcrypt==4.1.2
passlib==1.7.4
sqlalchemy==2.0.29
psycopg2==2.9.9
pydantic-settings==2.0.3
python-dotenv==1.0.0